Move weapon definition
