+++ src/definitions/weapons/definitions/alyss.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const alyss = {
   id: 'Alyss',
@@ -30,4 +30,4 @@ export const alyss = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
