+++ src/definitions/weapons/definitions/anka.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const anka = {
   id: 'Anka',
@@ -43,4 +43,4 @@ export const anka = {
   discharges: [],
   buffs: [],
   resources: [],
-} as const satisfies WeaponDefinition;
+} as const satisfies Weapon;
