+++ src/definitions/weapons/definitions/annabella.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const annabella = {
   id: 'Annabella',
@@ -53,4 +53,4 @@ export const annabella = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
