+++ src/definitions/weapons/definitions/asuka.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const asuka = {
   id: 'Asuka',
@@ -43,4 +43,4 @@ export const asuka = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
