+++ src/definitions/weapons/definitions/claudia.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const claudia = {
   id: 'Claudia',
@@ -31,4 +31,4 @@ export const claudia = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
