+++ src/definitions/weapons/definitions/cobalt-b.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const cobaltB = {
   id: 'Cobalt-B',
@@ -30,4 +30,4 @@ export const cobaltB = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
