+++ src/definitions/weapons/definitions/cocoritter.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const cocoritter = {
   id: 'Cocoritter',
@@ -43,4 +43,4 @@ export const cocoritter = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
