+++ src/definitions/weapons/definitions/fei-se.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const feiSe = {
   id: 'Fei Se',
@@ -30,4 +30,4 @@ export const feiSe = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
