+++ src/definitions/weapons/definitions/fenrir.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const fenrir = {
   id: 'Fenrir',
@@ -53,4 +53,4 @@ export const fenrir = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
