+++ src/definitions/weapons/definitions/fiona.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const fiona = {
   id: 'Fiona',
@@ -42,4 +42,4 @@ export const fiona = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
