+++ src/definitions/weapons/definitions/frigg.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const frigg = {
   id: 'Frigg',
@@ -52,4 +52,4 @@ export const frigg = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
