+++ src/definitions/weapons/definitions/gnonno.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const gnonno = {
   id: 'Gnonno',
@@ -31,4 +31,4 @@ export const gnonno = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
