+++ src/definitions/weapons/definitions/icarus.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const icarus = {
   id: 'Icarus',
@@ -30,4 +30,4 @@ export const icarus = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
