+++ src/definitions/weapons/definitions/ji-yu.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const jiYu = {
   id: 'Ji Yu',
@@ -43,4 +43,4 @@ export const jiYu = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
