+++ src/definitions/weapons/definitions/ling-han.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const lingHan = {
   id: 'Ling Han',
@@ -41,4 +41,4 @@ export const lingHan = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
