+++ src/definitions/weapons/definitions/lyra.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const lyra = {
   id: 'Lyra',
@@ -44,4 +44,4 @@ export const lyra = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
