+++ src/definitions/weapons/definitions/meryl.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const meryl = {
   id: 'Meryl',
@@ -17,4 +17,4 @@ export const meryl = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
