+++ src/definitions/weapons/definitions/mimi.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const mimi = {
   id: 'Huang (Mimi)',
@@ -30,4 +30,4 @@ export const mimi = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
