+++ src/definitions/weapons/definitions/ming-jing.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const mingJing = {
   id: 'Ming Jing',
@@ -43,4 +43,4 @@ export const mingJing = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
