+++ src/definitions/weapons/definitions/nan-yin.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const nanYin = {
   id: 'Nan Yin',
@@ -36,4 +36,4 @@ export const nanYin = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
