+++ src/definitions/weapons/definitions/nemesis.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const nemesis = {
   id: 'Nemesis',
@@ -65,4 +65,4 @@ export const nemesis = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
