+++ src/definitions/weapons/definitions/nola-flame-physical.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 import { nolaBase } from './nola-base';
 
 export const nolaFlamePhysical = {
@@ -46,4 +46,4 @@ export const nolaFlamePhysical = {
   discharges: [...nolaBase.discharges],
   buffs: [...nolaBase.buffs],
   resources: [...nolaBase.resources],
-} as const satisfies WeaponDefinition;
+} as const satisfies Weapon;
