+++ src/definitions/weapons/definitions/nola-physical-flame.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 import { nolaBase } from './nola-base';
 
 export const nolaPhysicalFlame = {
@@ -45,4 +45,4 @@ export const nolaPhysicalFlame = {
   discharges: [...nolaBase.discharges],
   buffs: [...nolaBase.buffs],
   resources: [...nolaBase.resources],
-} as const satisfies WeaponDefinition;
+} as const satisfies Weapon;
