+++ src/definitions/weapons/definitions/nola-volt-frost.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 import { nolaBase } from './nola-base';
 
 export const nolaVoltFrost = {
@@ -44,4 +44,4 @@ export const nolaVoltFrost = {
   discharges: [...nolaBase.discharges],
   buffs: [...nolaBase.buffs],
   resources: [...nolaBase.resources],
-} as const satisfies WeaponDefinition;
+} as const satisfies Weapon;
