+++ src/definitions/weapons/definitions/plotti.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const plotti = {
   id: 'Plotti',
@@ -43,4 +43,4 @@ export const plotti = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
