+++ src/definitions/weapons/definitions/roslyn.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const roslyn = {
   id: 'Roslyn',
@@ -56,4 +56,4 @@ export const roslyn = {
   discharges: [],
   buffs: [],
   resources: [],
-} as const satisfies WeaponDefinition;
+} as const satisfies Weapon;
