+++ src/definitions/weapons/definitions/saki-fuwa.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const sakiFuwa = {
   id: 'Saki Fuwa',
@@ -30,4 +30,4 @@ export const sakiFuwa = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
