+++ src/definitions/weapons/definitions/shiro.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const shiro = {
   id: 'Shiro',
@@ -17,4 +17,4 @@ export const shiro = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
