+++ src/definitions/weapons/definitions/tian-lang.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const tianLang = {
   id: 'Tian Lang',
@@ -43,4 +43,4 @@ export const tianLang = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
