+++ src/definitions/weapons/definitions/tsubasa.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const tsubasa = {
   id: 'Tsubasa',
@@ -17,4 +17,4 @@ export const tsubasa = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
