+++ src/definitions/weapons/definitions/yan-miao.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const yanMiao = {
   id: 'Yan Miao',
@@ -43,4 +43,4 @@ export const yanMiao = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
