+++ src/definitions/weapons/definitions/yanuo.ts
@@ -1,4 +1,4 @@
-import type { WeaponDefinition } from '../../../models/weapon-definition';
+import type { Weapon } from '../../types/weapon/weapon';
 
 export const yanuo = {
   id: 'Yanuo',
@@ -42,4 +42,4 @@ export const yanuo = {
   discharges: [],
   buffs: [],
   resources: [],
-} satisfies WeaponDefinition;
+} satisfies Weapon;
